+++ mirrord/config/src/target.rs
@@ -28,9 +28,13 @@ use crate::{
 #[derive(Deserialize, PartialEq, Eq, Clone, Debug, JsonSchema)]
 #[serde(untagged, rename_all = "lowercase")]
 pub enum TargetFileConfig {
+    // Generated when the value of the `target` field is a string, or when there is no target.
     // we need default else target value will be required in some scenarios.
     Simple(#[serde(default, deserialize_with = "string_or_struct_option")] Option<Target>),
     Advanced {
+        // Path is optional so that it can also be specified via env var instead of via conf file,
+        // but it is not optional in a resulting [`TargetConfig`] object - either there is a path,
+        // or the target configuration is `None`.
         #[serde(default, deserialize_with = "string_or_struct_option")]
         path: Option<Target>,
         namespace: Option<String>,
@@ -40,7 +44,7 @@ pub enum TargetFileConfig {
 #[derive(Serialize, Deserialize, Clone, Eq, PartialEq, Hash, Debug)]
 #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
 pub struct TargetConfig {
-    pub path: Option<Target>,
+    pub path: Target,
     pub namespace: Option<String>,
 }
 
@@ -54,33 +58,83 @@ impl FromMirrordConfig for TargetConfig {
     type Generator = TargetFileConfig;
 }
 
+impl TargetFileConfig {
+    /// Get the final path.
+    /// Will return the environment variable's value if set, if not the value from the
+    /// configuration (passed argument), and if that is not set as well, `None`.
+    ///
+    /// # Arguments
+    ///
+    /// * `path_from_config_file` - The optional value read from the config file.
+    fn get_optional_path(path_from_config_file: Option<Target>) -> Result<Option<Target>> {
+        FromEnvWithError::new("MIRRORD_IMPERSONATED_TARGET")
+            .or(path_from_config_file)
+            .source_value()
+            .transpose()
+    }
+
+    /// Get the final namespace.
+    /// Will return the environment variable's value if set, if not the value from the
+    /// configuration (passed argument), and if that is not set as well, `None`.
+    ///
+    /// # Arguments
+    ///
+    /// * `namespace_from_config_file` - The optional value read from the config file.
+    fn get_optional_namespace(
+        namespace_from_config_file: Option<String>,
+    ) -> Result<Option<String>> {
+        FromEnv::new("MIRRORD_TARGET_NAMESPACE")
+            .or(namespace_from_config_file)
+            .source_value()
+            .transpose()
+    }
+
+    /// Take the final values (after taking into account the values from the file and from env), and
+    /// return the optional `TargetConfig` (`None` if targetless).
+    ///
+    /// # Errors
+    /// * `ConfigError::TargetNamespaceWithoutTarget` - if namespace is some but path is `None`,
+    ///   because the target namespace does not mean anything without a target path. The user might
+    ///   have meant the agent namespace.
+    fn from_final_path_and_namespace(
+        path: Option<Target>,
+        namespace: Option<String>,
+    ) -> Result<Option<TargetConfig>> {
+        if let Some(path) = path {
+            Ok(Some(TargetConfig { path, namespace }))
+        } else if namespace.is_some() {
+            Err(ConfigError::TargetNamespaceWithoutTarget)
+        } else {
+            Ok(None)
+        }
+    }
+}
+
 impl MirrordConfig for TargetFileConfig {
-    type Generated = TargetConfig;
+    type Generated = Option<TargetConfig>;
 
+    /// Generate the final config object, out of the configuration parsed from a configuration file,
+    /// factoring in environment variables (which are also set by the front end - CLI/IDE-plugin).
+    ///
+    /// `None` if no target specified.
+    /// Specifying target namespace without target is not allowed and results in an error that
+    /// explains to the user what to do instead.
     fn generate_config(self) -> Result<Self::Generated> {
-        let config = match self {
-            TargetFileConfig::Simple(path) => TargetConfig {
-                path: FromEnvWithError::new("MIRRORD_IMPERSONATED_TARGET")
-                    .or(path)
+        match self {
+            TargetFileConfig::Simple(path) => {
+                // Namespace was not specified via file, get it from env var if set.
+                let namespace: Option<String> = FromEnv::new("MIRRORD_TARGET_NAMESPACE")
                     .source_value()
-                    .transpose()?,
-                namespace: FromEnv::new("MIRRORD_TARGET_NAMESPACE")
-                    .source_value()
-                    .transpose()?,
-            },
-            TargetFileConfig::Advanced { path, namespace } => TargetConfig {
-                path: FromEnvWithError::new("MIRRORD_IMPERSONATED_TARGET")
-                    .or(path)
-                    .source_value()
-                    .transpose()?,
-                namespace: FromEnv::new("MIRRORD_TARGET_NAMESPACE")
-                    .or(namespace)
-                    .source_value()
-                    .transpose()?,
-            },
-        };
-
-        Ok(config)
+                    .transpose()?;
+                let path = Self::get_optional_path(path)?;
+                Self::from_final_path_and_namespace(path, namespace)
+            }
+            TargetFileConfig::Advanced { path, namespace } => {
+                let path = Self::get_optional_path(path)?;
+                let namespace = Self::get_optional_namespace(namespace)?;
+                Self::from_final_path_and_namespace(path, namespace)
+            }
+        }
     }
 }
 
@@ -146,7 +200,7 @@ impl FromStr for Target {
             }
             Some("pod") => PodTarget::from_split(&mut split).map(Target::Pod),
             _ => Err(ConfigError::InvalidTarget(format!(
-                "Provided target: {target:?} is neither a pod or a deployment. Did you mean pod/{target:?} or deployment/{target:?}\n{FAIL_PARSE_DEPLOYMENT_OR_POD}",
+                "Provided target: {target} is neither a pod or a deployment. Did you mean pod/{target} or deployment/{target}\n{FAIL_PARSE_DEPLOYMENT_OR_POD}",
             ))),
         }
     }
@@ -218,21 +272,51 @@ mod tests {
     use crate::{config::MirrordConfig, util::testing::with_env_vars};
 
     #[rstest]
+    #[case(None, None, None)] // Nothing specified - no target config (targetless mode).
+    #[should_panic]
+    #[case(None, Some("ns"), None)] // Namespace without target - error.
+    #[case(
+        Some("pod/foo"),
+        None,
+        Some(TargetConfig{
+            path: Target::Pod(PodTarget {pod: "foo".to_string(), container: None}),
+            namespace: None
+        })
+    )] // Only pod specified
+    #[case(
+        Some("pod/foo/container/bar"),
+        None,
+        Some(TargetConfig{
+            path: Target::Pod(PodTarget {
+                pod: "foo".to_string(),
+                container: Some("bar".to_string())
+            }),
+            namespace: None
+        })
+    )] // Pod and container specified.
+    #[case(
+        Some("pod/foo"),
+        Some("baz"),
+        Some(TargetConfig{
+            path: Target::Pod(PodTarget {pod: "foo".to_string(), container: None}),
+            namespace: Some("baz".to_string())
+        })
+    )] // Pod and namespace specified.
     fn default(
-        #[values((None, None), (Some("pod/foobar"), Some(Target::Pod(PodTarget { pod: "foobar".to_string(), container: None }))))]
-        path: (Option<&str>, Option<Target>),
-        #[values((None, None), (Some("foo"), Some("foo")))] namespace: (Option<&str>, Option<&str>),
+        #[case] path_env: Option<&str>,
+        #[case] namespace_env: Option<&str>,
+        #[case] expected_target_config: Option<TargetConfig>,
     ) {
         with_env_vars(
             vec![
-                ("MIRRORD_IMPERSONATED_TARGET", path.0),
-                ("MIRRORD_TARGET_NAMESPACE", namespace.0),
+                ("MIRRORD_IMPERSONATED_TARGET", path_env),
+                ("MIRRORD_TARGET_NAMESPACE", namespace_env),
             ],
             || {
-                let target = TargetFileConfig::default().generate_config().unwrap();
+                let generated_target_config =
+                    TargetFileConfig::default().generate_config().unwrap();
 
-                assert_eq!(target.path, path.1);
-                assert_eq!(target.namespace.as_deref(), namespace.1);
+                assert_eq!(expected_target_config, generated_target_config);
             },
         );
     }